remove bokeh slider for compatible issue(TODO:FIX)
